Modify print job fields to include owner's country
As the queue should show the country instead of usernames. We use a
simple string field for this as using the country model in `trans`
would introduce cyclic dependency (and users can't change country
anyway).

A proper fix would be to move the country model to an app that both
`trans` and `print_job_queue` can use (`Translation`?) to avoid cyclic
dependency.